refactor: RedisMusicalResponse 클래스를 record 타입으로 변경
