+++ src/main/java/com/bootcamp/savemypodo/controller/AuthController.java
@@ -1,12 +1,6 @@
 package com.bootcamp.savemypodo.controller;
 
-import com.bootcamp.savemypodo.config.jwt.JwtTokenProvider;
-import com.bootcamp.savemypodo.entity.User;
-import com.bootcamp.savemypodo.global.exception.ErrorCode;
-import com.bootcamp.savemypodo.global.exception.UserException;
-import com.bootcamp.savemypodo.repository.UserRepository;
-import com.bootcamp.savemypodo.service.AuthService;
-import com.bootcamp.savemypodo.service.UserService;
+import com.bootcamp.savemypodo.service.auth.AuthService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -15,8 +9,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Arrays;
 
 @RestController
 @RequestMapping("/api/auth")
